+++ app/src/main/java/com/ebay/roy/weatherapp/model/Wind.java
@@ -7,7 +7,7 @@ import com.google.gson.annotations.SerializedName;
 public class Wind {
 
     private Double speed;
-    private Integer deg;
+    private Double deg;
 
     /**
      * 
@@ -32,7 +32,7 @@ public class Wind {
      * @return
      *     The deg
      */
-    public Integer getDeg() {
+    public Double getDeg() {
         return deg;
     }
 
@@ -41,7 +41,7 @@ public class Wind {
      * @param deg
      *     The deg
      */
-    public void setDeg(Integer deg) {
+    public void setDeg(Double deg) {
         this.deg = deg;
     }
